+++ iam_robolib/src/skills/joint_pose_continuous_skill.cpp
@@ -57,7 +57,7 @@ void JointPoseContinuousSkill::execute_skill_on_franka(run_loop *run_loop, Frank
     time += period_in_seconds;
     current_skill_time += period_in_seconds;
     traj_generator->time_ = current_skill_time;
-    traj_generator->dt_ = static_cast<float>(period_in_seconds);
+    traj_generator->dt_ = static_cast<double>(period_in_seconds);
     traj_generator->get_next_step();
 
     log_counter += 1;
